Split tests into unit and e2e
