+++ e2etests/tests/testdata/test_encoding.json
@@ -1,0 +1,8 @@
+{
+  "clknblocks": [
+    ["PNfT5qAAlAgFyQyoUhokyohywAAOYMJgdwPCRWBQOyCIsSEgePCo2CnRaON+FUog07AHTDUDARsUcJiSaYKNDiCAEeICbGYSZFhCVALQAylxDSAtSR4CsgiCmBjDiAEOGMfEi7ABkydqyKhIIFrBQQvFAUTDakAg0RyFYAWg8nE=", "1"],
+    ["WvjX1rZYiFCRQAyqptwAS7jAaACEhFApFxB9RSeYPTBLkYkIUjKacAIleYLoNUooU/AHYCQ2FSkUYCzZMiClDIGEIBKQXKZAogCAUpvOKikUKUI+CRgI+AkgGsqAJYkdZcHEAAWDEidgHbYgIjBpMgpUU2wCwgCoXwCCWCVAsPM=", "1"],
+    ["UPLX+gCgkbnVQhg4wBA8wogCUNA0QEbhPyHBPCBAKwSOmzJFWDGqWAZY/NsQX01Aw7A1UWBKAt5GPJib4BpLSHKQBesZbjsjYhTAQBBDZ7uQHKGNyRhCMhiWsAjaVEBcQlVkiCHzMhEimEgERlB/AQvFBwRjQHRywRQQQUuggnw=", "1", "2"],
+    ["MNXW4D9Gg5EYCQ2AQqGhBuAK+BiElCKJjyF1HjObORkqIDFgUTBZXQgEI4LAUc1k8tmMTkhijMynMKa44DiEWhDUAtAEphwapBCAFIpQIGgUAoIt0ZACOoxBEyBIFKIHRChFARbRunMh/GRi4hCRB2ZEJYZS1gTgAcAGXQGD++0=", "2"]
+  ]
+}
