+++ tools/ci.yml
@@ -1,17 +1,14 @@
 version: '3.4'
 services:
 
-
   tests:
-    image: data61/anonlink-app:${TAG:-latest}
+    image: data61/anonlink-test:${TAG:-latest}
     environment:
       - SERVER=http://nginx:8851
       - INITIAL_DELAY=20
       - JAEGER_AGENT_HOST=jaeger
-    command: /bin/sh -c "dockerize -wait tcp://nginx:8851/v1/status -timeout 1m python -m pytest -n 4 entityservice/tests --junitxml=testResults.xml -x"
     depends_on:
-      - db
-      - redis
+      - backend
       - worker
       - nginx
